==> Utah accidents analysis/03 Assign crashes to road segments.ipynb <==
### Connect to GIS

from arcgis import GIS
from arcgis.features import SpatialDataFrame, FeatureLayer

gis = GIS(profile='deldev')

### Load crash data

import pandas as pd

pd.set_option('display.height', 1000)
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

df = pd.read_pickle('df_utah.pkl')

df.head()

df.shape

df[df.RAMPID.notnull()].shape

df['RouteName'] = df['RouteName'].apply(lambda x: x.zfill(5)) # mislabeled, zero pad so it matches LRS names
df.loc[df.RouteName == '0011P', 'RouteName'] = '089AP'        # mislabeled - QA from ArcGIS Pro inspection

route_names = df.RouteName.unique()
len(route_names)

df[df.RouteName.str.len()<5]

### Make LABEL column to join with Safety Index Layer

The RouteName contains just the road names. If the crash occurred on a ramp or connector, RAMPID is a number and the ramp/connector label is in the project column. The LABEL column reads these from the RouteName or project column conditionally. In some cases, RAMPID is a number but project does not include the ramp label (which is supposed to be 11 chars long). Handle that as well.

df['LABEL'] = df.project.where(df.RAMPID.notnull() & (df.project.str.len()==11), df['RouteName'])

df[df.LABEL.str.len()==11].shape

df[df.LABEL.str.len()<11].head()

### Safety Index layer

from arcgis.features import SpatialDataFrame, FeatureLayer
safetyindex_lyr = FeatureLayer('https://maps.udot.utah.gov/arcgis/rest/services/Safety_SafetyIndex/MapServer/0')
safetyindex_df = SpatialDataFrame.from_layer(safetyindex_lyr)

safetyindex_df.head()

safetyindex_df.shape

### Choosing which data points to use

crashes_ramps = df[df['LABEL'].str.len()==11]  # crashes on ramps

crashes_roads = df[df['LABEL'].str.len()!=11]  # crashes on roads

len(crashes_ramps.LABEL.unique())

len(crashes_roads.LABEL.unique())

len(safetyindex_df.LABEL.unique())   # 2099 - 904 roads will not be matched, hopefully they're minor roads

len(crashes_roads[crashes_roads['LABEL'].str.len()>5]['LABEL'].unique())

len(crashes_roads[crashes_roads['LABEL'].str.len()==5]['LABEL'].unique())

crashes_roads[crashes_roads['LABEL'].str.len()<5]['LABEL']

len(crashes_ramps), len(crashes_roads)

crashes_roads[crashes_roads.LABEL.isin(safetyindex_df.LABEL.unique())].shape

len(crashes_ramps)

342376/414637          # data points ignoring ramps, have all data

(342376+16108)/414637  # data points including ramps, missing ramp specific data from safety index layer, can use road data

372524/414637          # data points if using LRS layer, calculated in another notebook - missing safety index data

We'll not be able to use approx 3% of the crash data if we use safety index layer instead of LRS layer. However, LRS layer is missing the data we want to use (AADT, safety rating, crashes per mile, etc). Proceeding with Safety Index layer.

crashes_roads_segments = pd.merge(crashes_roads, safetyindex_df, on='LABEL')

crashes_roads_segments.shape

crashes_roads_segments = crashes_roads_segments[crashes_roads_segments.BeginMilePost.between(crashes_roads_segments.BEG_MP, crashes_roads_segments.END_MP)]

crashes_roads_segments.shape

crashes_roads_segments.head()

crashes_roads_segments.to_pickle('crashes_roads_segments.pkl')

crashes_roads_segments.to_csv('crashes_roads_segments.csv')

safetyindex_df.to_csv('safetyindex_roads.csv')
