==> Utah accidents analysis/05 Fetch Hourly Weather.ipynb <==
import pandas as pd
import pandas as pd
import numpy as np
from tqdm import tqdm
from MesoPy import Meso

pd.set_option('display.height', 1000)
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

active_stn = pd.read_pickle('HourlyWeatherActiveStations.pkl')

!mkdir Weather

m = Meso(token='...')

for year in range(2010, 2011):
    print(year)
    for stn in tqdm(active_stn['Station ID']):
        try:
            data = m.timeseries(start='{}01010000'.format(year), end='{}12312300'.format(year), stid=stn)
            obs = data['STATION'][0]['OBSERVATIONS']
            obs_df = pd.DataFrame(obs)
            obs_df['year'] = year
            obs_df['Station ID'] = stn
            obs_df.to_pickle('Weather/{}_{}.pkl'.format(stn, year))
        except:
            print('Error getting {} data for {}'.format(stn, year))

for year in range(2011, 2012):
    print(year)
    for stn in tqdm(active_stn['Station ID']):
        try:
            data = m.timeseries(start='{}01010000'.format(year), end='{}12312300'.format(year), stid=stn)
            obs = data['STATION'][0]['OBSERVATIONS']
            obs_df = pd.DataFrame(obs)
            obs_df['year'] = year
            obs_df['Station ID'] = stn
            obs_df.to_pickle('Weather/{}_{}.pkl'.format(stn, year))
        except:
            print('Error getting {} data for {}'.format(stn, year))
