==> Utah accidents analysis/06 Combine Annual Weather.ipynb <==
import pandas as pd
from tqdm import tqdm

def chunks(l, n):
    """Yield successive n-sized chunks from l."""
    for i in range(0, len(l), n):
        yield l[i:i + n]

for year  in range(2010, 2018):
    print(year)
    files = glob('Weather/*{}.pkl'.format(year))
    
    combined = pd.DataFrame()

    for chunk in tqdm(chunks(files, 20)):
        wdfs = [pd.read_pickle(file) for file in chunk]
        combined = combined.append(wdfs, ignore_index=True)
        
    combined.to_pickle('hourly_weather_{}.pkl'.format(year))
