==> Utah accidents analysis/07 Find nearest weather station to crash locations.ipynb <==
import pandas as pd
import numpy as np

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

crashes = pd.read_pickle('crashes_roads_segments.pkl')

crashes.head()

weatherstn = pd.read_pickle('HourlyWeatherActiveStations.pkl')
weatherstn.head()

weatherstn.dtypes

cols = ['Latitude', 'Longitude']
weatherstn[cols] = weatherstn[cols].apply(pd.to_numeric, errors='coerce')

weather_coords = weatherstn[['Latitude', 'Longitude']].values[np.newaxis, :]           # airport=weather, places = crashes
crashes_coords = np.rollaxis(crashes[['lat', 'lon']].values[np.newaxis, :], 0, -1)

dist_coords = ((crashes_coords - weather_coords)**2).sum(axis=-1)
min_coords = dist_coords.argmin(axis=1)

crashes_coords.shape, weather_coords.shape, dist_coords.shape, min_coords.shape

merge_data = crashes.copy()

print(merge_data.shape)

merge_data['weatherstn_index'] = weatherstn.index[min_coords]

df = weatherstn.loc[merge_data.weatherstn_index, ['Station ID', 'Station Name']]
merge_data[['Station ID', 'Station Name']] = df.set_index(merge_data.index)

print(merge_data.shape)

merge_data.head()

merge_data.to_pickle('crashes_weatherstn.pkl')
