==> Utah accidents analysis/09 Merge weather and create yearly samples.ipynb <==
import pandas as pd
import pandas as pd
import numpy as np
from tqdm import tqdm
from glob import glob

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

pd.options.mode.chained_assignment = None 

crashes = pd.read_pickle('crashes_weatherstn.pkl')

year = 2010

positive = crashes[(crashes.year == year)]
negative = pd.read_pickle('negative_{}.pkl'.format(year))
positive['class'] = 1
negative['class'] = 0
positive.shape, negative.shape

samples = pd.concat([positive, negative], ignore_index=True)
samples.shape

samples.head()

samples.day = samples.day.astype(int)
samples.dayofweek = samples.dayofweek.astype(int)
samples.hour = samples.hour.astype(int)
samples.month = samples.month.astype(int)
samples.year = samples.year.astype(int)
samples.dayofyear = samples.dayofyear.astype(int)

samples['TOT_C'] = samples['TOT_C'].astype(int)
samples['REGION'] = samples['REGION'].astype(int)
samples['SAF_INDEX'] = samples['SAF_INDEX'].astype(int)
samples['TOT_SEV_C'] = samples['TOT_SEV_C'].astype(int)
samples['CrashSeverity'] = samples['CrashSeverity'].astype(int)
samples['NUMBERVEHICLESINVOLVED'] = samples['NUMBERVEHICLESINVOLVED'].astype(int)

samples = samples.sample(frac=1).reset_index(drop=True)

samples.head()

weather = pd.read_pickle('hourly_weather_{}.pkl'.format(year))

weather['datetime'] = pd.to_datetime(weather.date_time, infer_datetime_format=True)

weather['dayofyear'] = weather.datetime.dt.dayofyear
weather['hour'] = weather.datetime.dt.hour

weather['key'] = weather['year'].map(str) + \
                 weather['dayofyear'].map(str) + \
                 weather['hour'].map(str) + weather['Station ID']

samples['key'] = samples['year'].map(str) + \
                      samples['dayofyear'].map(str) + \
                      samples['hour'].map(str) + samples['Station ID']

samples.shape

len(weather.key.unique())

len(weather)

unique_weather = weather.drop_duplicates(subset='key')

samples_weather = pd.merge(samples, unique_weather, on='key', how='left')

samples_weather.shape

samples_weather.head()

samples_weather.to_pickle('samples_{}.pkl'.format(year))

!ls -l samples*.pkl

!df -h

!tar cvf samples.tar samples_*.pkl

!gzip samples.tar

from IPython.display import FileLink

FileLink('samples.tar.gz')
